Update build dependencies, refactor code and bump version to 2.0 snapshot
This commit includes a large number of changes resulting in a complete break with previous versions, thereby warranting the update to version 2.0 snapshot. A significant modification is bringing the code into compliance with boot version 3.0-SNAPSHOT. Additionally, several changes are implemented in the build dependencies. The `load.kt` file is renamed to `loadMinecraft.kt` and some test cases and functionalities in different files are updated or removed for efficiency. Changes also include increasing concurrency and improving downloads of Minecraft assets and libraries. The commit provides a solid base for further improvements and feature additions.